Remove the Select overlay template in favor of a patch
RT 5 switches to bootstrap, so the Select widget is now different
between RT 4 and RT 5. This makes it harder to provide an overlay
that will work on both.

Multiple support has been in core RT for a long time, so move
the update that added it to a patch for continuity on older
RTs and remove the overlay.